CURA-48 Moved fetchPatients function from PatientList to Patients in frontend.
