+++ frontend/src/components/PatientList.jsx
@@ -1,25 +1,4 @@
-import { useState, useEffect } from 'react';
-import { useAuth } from '../context/AuthContext';
-import axiosInstance from '../axiosConfig';
-
-const PatientList = () => {
-    const { user } = useAuth();
-    const [patients, setPatients] = useState([]);
-
-    useEffect(() => {
-        const fetchPatients = async () => {
-            try {
-                const response = await axiosInstance.get('/api/patients/', {
-                    headers: { Authorization: `Bearer ${user.token}` },
-                });
-                setPatients(response.data);
-            } catch (error) {
-                alert('Error: Failed to fetch patient.')
-            }
-        };
-        if (user) fetchPatients();
-    }, [user]);
-
+const PatientList = ({ patients }) => {
     return (
         <div className="bg-white mb-6 p-6 rounded shadow-md">
             <h1 className="font-bold mb-4 text-2xl">Patient List</h1>
