+++ frontend/src/pages/Patients.jsx
@@ -1,11 +1,33 @@
+import { useState, useEffect } from 'react';
+import { useAuth } from '../context/AuthContext';
+import axiosInstance from '../axiosConfig';
 import PatientForm from "../components/PatientForm";
 import PatientList from "../components/PatientList";
 
 const Patients = () => {
+    const { user } = useAuth();
+    const [patients, setPatients] = useState([]);
+
+    useEffect(() => {
+        const fetchPatients = async () => {
+            try {
+                const response = await axiosInstance.get('/api/patients/', {
+                    headers: { Authorization: `Bearer ${user.token}` },
+                });
+                setPatients(response.data);
+            } catch (error) {
+                alert('Error: Failed to fetch patient.')
+            }
+        };
+        if (user) fetchPatients();
+    }, [user]);
+
     return (
         <div className="container mx-auto p-6">
             <PatientForm/>
-            <PatientList/>
+            <PatientList
+                patients={patients}
+            />
         </div>
     );
 };
